introcartoon work

+++ Classes/intro/IntroAnimFrame2.h
@@ -1,7 +1,7 @@
 #import "IntroAnim.h"
 
 @interface IntroAnimFrame2 : IntroAnimFrame {
-	int ct;
+	float ct;
 	CCSprite *bg;
 	CCSprite *ground;
 	CCSprite *robot1, *robot2, *robot3;
